update COSMO calc

+++ scripts/calculation/main_COSMO_calc.py
@@ -110,7 +110,8 @@ for mol_id, smi in mol_ids_smis[args.task_id::args.num_tasks]:
         suboutputs_dir = os.path.join(outputs_dir, f"outputs_{ids}")
         os.makedirs(suboutputs_dir, exist_ok=True)
         mol_id_path = os.path.join(subinputs_dir, f"{mol_id}.in")
-        if not os.path.exists(os.path.join(suboutputs_dir, f"{mol_id}.tar")) and not os.path.exists(mol_id_path):
+        tmp_mol_id_path = os.path.join(subinputs_dir, f"{mol_id}.tmp")
+        if not os.path.exists(os.path.join(suboutputs_dir, f"{mol_id}.tar")) and not os.path.exists(mol_id_path) and not os.path.exists(tmp_mol_id_path):
             os.makedirs(subinputs_dir, exist_ok=True)
             with open(mol_id_path, "w+") as f:
                 f.write(mol_id)
